+++ ICD.Connect.Misc.CrestronPro/Devices/Io/CenIo/AbstractCenIoAdapterSettings.cs
@@ -1,0 +1,37 @@
+using ICD.Common.Utils;
+using ICD.Common.Utils.Xml;
+using ICD.Connect.Devices;
+using ICD.Connect.Settings.Attributes.SettingsProperties;
+
+namespace ICD.Connect.Misc.CrestronPro.Devices.Io.CenIo
+{
+	public abstract class AbstractCenIoAdapterSettings : AbstractDeviceSettings, ICenIoAdapterSettings
+	{
+		private const string IPID_ELEMENT = "IPID";
+
+		[CrestronByteSettingsProperty]
+		public byte? Ipid { get; set; }
+
+		/// <summary>
+		/// Writes property elements to xml.
+		/// </summary>
+		/// <param name="writer"></param>
+		protected override void WriteElements(IcdXmlTextWriter writer)
+		{
+			base.WriteElements(writer);
+
+			writer.WriteElementString(IPID_ELEMENT, Ipid == null ? null : StringUtils.ToIpIdString(Ipid.Value));
+		}
+
+		/// <summary>
+		/// Updates the settings from xml.
+		/// </summary>
+		/// <param name="xml"></param>
+		public override void ParseXml(string xml)
+		{
+			base.ParseXml(xml);
+
+			Ipid = XmlUtils.TryReadChildElementContentAsByte(xml, IPID_ELEMENT) ?? 0;
+		}
+	}
+}