+++ ICD.Connect.Misc.CrestronPro/Devices/Io/CenIo/ICenIoAdapter.cs
@@ -1,0 +1,8 @@
+using ICD.Connect.Devices;
+
+namespace ICD.Connect.Misc.CrestronPro.Devices.Io.CenIo
+{
+	public interface ICenIoAdapter : IDevice, IPortParent
+	{
+	}
+}