+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added Crestron Ethernet IO devices CEN-IO-COM-102, CEN-IO-DI-104, CEN-IO-IR-104, CEN-IO-RY-104
 
 ## [7.1.0] - 2019-01-29
 ### Added